api/requirements.txt
fastapi
uvicorn
groq
motor
dnspython
pypdf
python-multipart
python-dotenv